move userspace interface from video.h to sys/videoio.h
