+++ drivers/video/video.c
@@ -36,6 +36,7 @@
 #include <nuttx/kmalloc.h>
 #include <nuttx/mutex.h>
 
+#include <nuttx/video/video.h>
 #include <nuttx/video/imgsensor.h>
 #include <nuttx/video/imgdata.h>
