+++ drivers/video/video_framebuff.h
@@ -25,7 +25,7 @@
  * Included Files
  ****************************************************************************/
 
-#include <nuttx/video/video.h>
+#include <sys/videoio.h>
 #include <nuttx/mutex.h>
 
 /****************************************************************************
